Dry run for patch and check for symbolic link.

+++ install.sh
@@ -2,11 +2,11 @@
 
 TARGET_DIR="/usr/share/huya/xkb"
 if [ ! -d $TARGET_DIR ]; then
-    echo "Installing HUYA/XKB for the first time"
+  echo "Installing HUYA/XKB for the first time"
 else
-    echo "HUYA/XKB is already installed, reinstalling"
-    cp $TARGET_DIR/*.bup /tmp
-    rm -rf $TARGET_DIR
+  echo "HUYA/XKB is already installed, reinstalling"
+  cp $TARGET_DIR/*.bup /tmp
+  rm -rf $TARGET_DIR
 fi
 
 echo "Cloning the repository..."
@@ -14,10 +14,13 @@ git clone https://github.com/mudasobwa/huya-xkb.git $TARGET_DIR
 cd $TARGET_DIR
 [ "$1" = "ask" ] && export ASK="true"
 cd /usr/share/X11/xkb/symbols || exit 1
-ln -s $TARGET_DIR/ru_es.layout ru_es
+if [ ! -f ru_es ] ; then
+  ln -s $TARGET_DIR/ru_es.layout ru_es
+fi
 for i in ru es ; do
   cp $i $TARGET_DIR/$i.bup
-  patch $i < $TARGET_DIR/$i.patch
+  patch --dry-run $i < $TARGET_DIR/$i.patch
+  patch -b $i < $TARGET_DIR/$i.patch
 done
 
 echo "Removing xkb cache..."
